moved assets into their own folder
