+++ src/pages/PassportRedemption/LyftPromo.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import { TitleRow, SubTitle, Button, CardContainer } from './style';
-import CitiBikeImg from './citi_bike.png';
+import CitiBikeImg from './Assets/citi_bike.png';
 import { useTranslation } from 'react-i18next';
 import CopyToClipboard from 'react-copy-to-clipboard';
 import Tooltip from 'react-bootstrap/Tooltip';
