+++ src/pages/PassportRedemption/RedemptionSelectScreen.tsx
@@ -9,7 +9,7 @@ import {
   DefaultFooter,
 } from './RedemptionFooters';
 import { CardText } from './style';
-import CircleLogo from './CircleLogo.png';
+import CircleLogo from './Assets/CircleLogo.png';
 
 import {
   getPassportTickets,
